refactor: move experiment detail specific components to page dir
