+++ webui/react/src/pages/ExperimentDetails/ExperimentInfoBox.module.scss
